Add a way for the flagship app to create session
The flagship app, with its maximal permission, can create session codes
by calling the new endpoint /auth/session_code. Then, it can put this
session code in the query string of a webapp URL, and open this URL in a
webview. A new session will automatically be created for the webview, to
avoid asking the user to type their password again.

As a security, a session code can only be used once.